Move headless classes into headless subsystem
