+++ engine/src/main/java/org/terasology/engine/subsystem/headless/HeadlessGraphics.java
@@ -25,12 +25,19 @@ import org.terasology.engine.modes.GameState;
 import org.terasology.engine.subsystem.DisplayDevice;
 import org.terasology.engine.subsystem.EngineSubsystem;
 import org.terasology.engine.subsystem.RenderingSubsystemFactory;
+import org.terasology.engine.subsystem.headless.assets.HeadlessFont;
+import org.terasology.engine.subsystem.headless.assets.HeadlessMaterial;
+import org.terasology.engine.subsystem.headless.assets.HeadlessMesh;
+import org.terasology.engine.subsystem.headless.assets.HeadlessShader;
+import org.terasology.engine.subsystem.headless.assets.HeadlessSkeletalMesh;
+import org.terasology.engine.subsystem.headless.assets.HeadlessTexture;
+import org.terasology.engine.subsystem.headless.device.HeadlessDisplayDevice;
+import org.terasology.engine.subsystem.headless.renderer.GUIManagerHeadless;
 import org.terasology.engine.subsystem.headless.renderer.HeadlessRenderingSubsystemFactory;
+import org.terasology.engine.subsystem.headless.renderer.ShaderManagerHeadless;
 import org.terasology.logic.manager.GUIManager;
-import org.terasology.logic.manager.GUIManagerHeadless;
 import org.terasology.registry.CoreRegistry;
 import org.terasology.rendering.ShaderManager;
-import org.terasology.rendering.ShaderManagerHeadless;
 import org.terasology.rendering.VertexBufferObjectManager;
 import org.terasology.rendering.assets.animation.MeshAnimation;
 import org.terasology.rendering.assets.animation.MeshAnimationData;
@@ -53,12 +60,6 @@ import org.terasology.rendering.assets.texture.TextureData;
 import org.terasology.rendering.assets.texture.subtexture.Subtexture;
 import org.terasology.rendering.assets.texture.subtexture.SubtextureData;
 import org.terasology.rendering.assets.texture.subtexture.SubtextureFromAtlasResolver;
-import org.terasology.rendering.headless.HeadlessFont;
-import org.terasology.rendering.headless.HeadlessMaterial;
-import org.terasology.rendering.headless.HeadlessMesh;
-import org.terasology.rendering.headless.HeadlessShader;
-import org.terasology.rendering.headless.HeadlessSkeletalMesh;
-import org.terasology.rendering.headless.HeadlessTexture;
 import org.terasology.rendering.iconmesh.IconMeshResolver;
 import org.terasology.rendering.nui.NUIManager;
 import org.terasology.rendering.nui.internal.LwjglCanvasRenderer;
