+++ engine/src/main/java/org/terasology/engine/subsystem/headless/assets/HeadlessFont.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 MovingBlocks
+ * Copyright 2014 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.engine.internal;
+package org.terasology.engine.subsystem.headless.assets;
 
-public final class TimeSystem extends TimeBase {
-    @Override
-    public long getRawTimeInMs() {
-        return System.nanoTime() / 1000000;
+import org.newdawn.slick.Color;
+import org.terasology.asset.AssetUri;
+import org.terasology.rendering.assets.font.BaseFont;
+import org.terasology.rendering.assets.font.FontData;
+
+public class HeadlessFont extends BaseFont {
+
+    public HeadlessFont(AssetUri uri, FontData data) {
+        super(uri, data);
     }
+
+    @Override
+    public void drawString(int x, int y, String text, Color color) {
+        // Do nothing
+    }
+
 }
