+++ engine/src/main/java/org/terasology/engine/subsystem/headless/assets/HeadlessMaterial.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.rendering.headless;
+package org.terasology.engine.subsystem.headless.assets;
 
 import java.nio.FloatBuffer;
