+++ engine/src/main/java/org/terasology/engine/subsystem/headless/device/HeadlessDisplayDevice.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.engine.subsystem.headless;
+package org.terasology.engine.subsystem.headless.device;
 
 import org.terasology.engine.subsystem.DisplayDevice;
