+++ engine/src/main/java/org/terasology/engine/subsystem/headless/renderer/ShaderManagerHeadless.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.rendering;
+package org.terasology.engine.subsystem.headless.renderer;
 
+import org.terasology.rendering.ShaderManager;
 import org.terasology.rendering.assets.material.Material;
 import org.terasology.rendering.assets.texture.Texture;
